refactor: add success toast
